Fix behaviour of metal when creating a new project
In this PR we apply a change due to Denis Bozzato to fix a bug when
using hpegl to create a new metal project.  This is the first in an
anticipated series of fixes to metal support.  With this change, the
call to RefreshResources when creating the metal client is only issued
if the project_id is not empty. In this way metal project creation will
succeed.